% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Bernoulli}
\alias{Bernoulli}
\title{Bernoulli Distribution (Categorical with k=1)}
\usage{
Bernoulli()
}
\value{
Bernoulli Distribution Object
}
\description{
Bernoulli Distribution (Categorical with k=1)
}
