% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRClassifier.R
\name{create_classifier}
\alias{create_classifier}
\title{Create a NGBoostClassifier object}
\usage{
create_classifier(
  Base = DecisionTreeRegressor(),
  Dist = Bernoulli(),
  col_sample = 1,
  learning_rate = 0.01,
  minibatch_frac = 1,
  n_estimators = as.integer(500),
  natural_gradient = TRUE,
  random_state = NULL,
  tol = 1e-04,
  verbose = TRUE,
  verbose_eval = as.integer(100)
)
}
\arguments{
\item{Dist}{Distribution should be used}

\item{col_sample}{the percent subsample of columns to use in each boosting iteration}

\item{learning_rate}{the learning rate}

\item{minibatch_frac}{the percent subsample of rows to use in each boosting iteration}

\item{n_estimators}{the number of boosting iterations to fit}

\item{natural_gradient}{logical flag indicating whether the natural gradient should be used}

\item{tol}{numerical tolerance to be used in optimization}

\item{verbose}{flag indicating whether output should be printed during fitting}

\item{verbose_eval}{increment (in boosting iterations) at which output should be printed}
}
\value{
A NGBoostRegressor object
}
\description{
Create a NGBoostClassifier object
}
