% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRRegressor.R
\name{create_regressor}
\alias{create_regressor}
\title{Create a NGBoostRegressor object}
\usage{
create_regressor(
  natural_gradient = TRUE,
  n_estimators = as.integer(500),
  learning_rate = 0.01,
  minibatch_frac = 1,
  col_sample = 1,
  verbose = TRUE,
  verbose_eval = as.integer(100),
  tol = 1e-04
)
}
\arguments{
\item{natural_gradient}{logical flag indicating whether the natural gradient should be used}

\item{n_estimators}{the number of boosting iterations to fit}

\item{learning_rate}{the learning rate}

\item{minibatch_frac}{the percent subsample of rows to use in each boosting iteration}

\item{col_sample}{the percent subsample of columns to use in each boosting iteration}

\item{verbose}{flag indicating whether output should be printed during fitting}

\item{verbose_eval}{increment (in boosting iterations) at which output should be printed}

\item{tol}{numerical tolerance to be used in optimization}
}
\value{
A NGBoostRegressor object
}
\description{
Create a NGBoostRegressor object
}
