% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{DecisionTreeRegressor}
\alias{DecisionTreeRegressor}
\title{DecisionTreeRegressor}
\usage{
DecisionTreeRegressor(
  criterion = "friedman_mse",
  min_samples_split = 2,
  min_samples_leaf = 1,
  min_weight_fraction_leaf = 0,
  max_depth = 3,
  splitter = "best",
  random_state = NULL
)
}
\arguments{
\item{criterion}{criterion}

\item{min_samples_split}{min_samples_split}

\item{min_samples_leaf}{min_samples_leaf}

\item{max_depth}{max_depth}

\item{splitter}{splitter}

\item{random_state}{random_state}
}
\value{
DecisionTreeRegressor Object
}
\description{
DecisionTreeRegressor
}
