% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{default_linear_learner}
\alias{default_linear_learner}
\title{Default LinearTreeRegressor}
\usage{
default_linear_learner()
}
\value{
Default LinearTreeRegressor
}
\description{
Default LinearTreeRegressor
}
