% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{default_tree_learner}
\alias{default_tree_learner}
\title{Default DecisionTreeRegressor}
\usage{
default_tree_learner()
}
\value{
Default DecisionTreeRegressor
}
\description{
Default DecisionTreeRegressor
}
