% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRClassifier.R
\name{fit_classifier}
\alias{fit_classifier}
\title{Train a NGBoostClassifier object}
\usage{
fit_classifier(ngbr_cla, X_train, Y_train, X_val = NULL, Y_val = NULL)
}
\arguments{
\item{X_train}{Training data (numeric). Is an object where samples are in rows and features are in columns.}

\item{Y_train}{Training data. A numeric (integer)  vector containing the outcome for each sample.}

\item{X_val}{validattion data (numeric). Is an object where samples are in rows and features are in columns.}

\item{Y_val}{Validation data. A numeric (integer)  vector containing the outcome for each sample.}

\item{ngbr_reg}{NGBoostRegressor obcjet}
}
\description{
Train a NGBoostClassifier object
}
