% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ngboost}
\alias{install_ngboost}
\title{Installation Python ngboost module}
\usage{
install_ngboost(envname = NULL, method = "auto", conda = "auto")
}
\arguments{
\item{envname}{The name, or full path, of the environment in which Python packages are to be installed. When NULL (the default), the active environment as set by the RETICULATE_PYTHON_ENV variable will be used; if that is unset, then the r-reticulate environment will be used.}

\item{method}{Installation method. By default, "auto" automatically finds a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows.}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate to automatically find an appropriate conda binary. See Finding Conda for more details.}
}
\description{
Installation Python ngboost module
}
