% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{k_categorical}
\alias{k_categorical}
\title{Categorical Distribution}
\usage{
k_categorical(K)
}
\arguments{
\item{K}{number of categories}
}
\value{
Categorical Distribution Object
}
\description{
Categorical Distribution
}
