% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{MultivariateNormal}
\alias{MultivariateNormal}
\title{Multivariante Normal Distribution}
\usage{
MultivariateNormal(k)
}
\arguments{
\item{k}{number of dimensions}
}
\value{
Multivariante Normal Distribution Object
}
\description{
Multivariante Normal Distribution
}
