% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ngboost}
\alias{ngboost}
\title{NGBoost for R}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
ngboost
}
\description{
\href{https://stanfordmlgroup.github.io}{NGBoost} bla bla bla
}
\keyword{datasets}
