% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ngboost_python_version}
\alias{ngboost_python_version}
\title{Version Number of the python module NGBoost}
\usage{
ngboost_python_version()
}
\value{
Version Number of the python module NGBoost
}
\description{
Version Number of the python module NGBoost
}
