% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRClassifier.R
\name{predict_classifier_dist}
\alias{predict_classifier_dist}
\title{Predict Distribution (Bernoulli o Cateogiral) using NGBoostClassifierr object}
\usage{
predict_classifier_dist(ngbr_cla, new_data)
}
\arguments{
\item{ngbr_reg}{NGBoostClassifier obcjet}

\item{new_Data}{Is an object where new samples are in rows and features are in columns.}
}
\value{
A list with loc(mean) and scale(desviation)
}
\description{
Predict Distribution (Bernoulli o Cateogiral) using NGBoostClassifierr object
}
