% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRClassifier.R
\name{predict_classifier_prob}
\alias{predict_classifier_prob}
\title{Predict using a NGBoostClassifier object as a probability}
\usage{
predict_classifier_prob(ngbr_cla, new_data)
}
\arguments{
\item{ngbr_cla}{NGBoostClassifier obcjet}

\item{new_Data}{Is an object where new samples are in rows and features are in columns.}
}
\value{
Predictions
}
\description{
Predict using a NGBoostClassifier object as a probability
}
