% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRRegressor.R
\name{predict_regressor}
\alias{predict_regressor}
\title{Predict using a NGBoostRegressor object}
\usage{
predict_regressor(ngbr_reg, new_data)
}
\arguments{
\item{ngbr_reg}{NGBoostRegressor obcjet}

\item{new_Data}{Is an object where new samples are in rows and features are in columns.}
}
\value{
Predictions
}
\description{
Predict using a NGBoostRegressor object
}
