% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRRegressor.R
\name{predict_regressor_dist}
\alias{predict_regressor_dist}
\title{Predict Normal distribution using NGBoostRegressor object}
\usage{
predict_regressor_dist(ngbr_reg, new_data)
}
\arguments{
\item{ngbr_reg}{NGBoostRegressor obcjet}
}
\value{
A list with loc(mean) and scale(desviation)
}
\description{
Predict Normal distribution using NGBoostRegressor object
}
