% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{Ridge}
\alias{Ridge}
\title{LinearTreeRegressor}
\usage{
Ridge(alpha = 0, random_state = NULL)
}
\arguments{
\item{alpha}{alpha}

\item{random_state}{random_state}
}
\value{
LinearTreeRegressor Object
}
\description{
LinearTreeRegressor
}
