% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRRegressor.R
\name{create_regressor}
\alias{create_regressor}
\title{Create a NGBoostRegressor object}
\usage{
create_regressor(
  natural_gradient = TRUE,
  n_estimators = as.integer(500),
  learning_rate = 0.01,
  minibatch_frac = 1,
  col_sample = 1,
  verbose = TRUE,
  verbose_eval = as.integer(100),
  tol = 1e-04
)
}
\arguments{
\item{natural_gradient}{boolean if the natural gradient is used}
}
\value{
A NGBoostRegressor object
}
\description{
Create a NGBoostRegressor object
}
