% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboostRRegressor.R
\name{fit_regressor}
\alias{fit_regressor}
\title{Train a NGBoostRegressor object}
\usage{
fit_regressor(ngbr_reg, X_train, Y_train, X_val, Y_val)
}
\arguments{
\item{ngbr_reg}{NGBoostRegressor obcjet}
}
\value{
Nothing
}
\description{
Train a NGBoostRegressor object
}
